♻️ [Refactor] 게시물 작성 화면 코드 분리
